backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
python-multipart==0.0.12
httpx==0.27.2
Pillow==10.4.0
bcrypt==4.1.2
PyJWT==2.8.0
